Refactor ExerciseAddModal
