+++ react-app/src/components/Exercises/ExerciseNavbar.js
@@ -1,15 +1,14 @@
 import React from "react";
 import { Navbar, Button, Container } from "react-bootstrap";
 import { FaPlus } from "react-icons/fa";
+import ExerciseAddModal from "./ExerciseAddModal";
 
 const ExerciseNavbar = ({ onAddExercise }) => {
   return (
     <Navbar variant="dark" bg="dark" expand="lg" className="secondary-navbar">
       <Container>
         <Navbar.Brand>Exercise Bank</Navbar.Brand>
-        <Button variant="outline-light" className="ms-3" onClick={onAddExercise}>
-            <FaPlus />
-          </Button>
+        <ExerciseAddModal/>
         <Navbar.Toggle aria-controls="navbarScroll" />
       </Container>
     </Navbar>
