+++ react-app/src/views/ExerciseManagement.js
@@ -1,21 +1,16 @@
 import React, { useState, useEffect } from "react";
 import { Container, Form, Button, Alert, Row, Modal } from "react-bootstrap";
-import ExerciseCard from "../components/ExerciseCard";
+import ExerciseCard from "../components/Exercises/ExerciseCard";
 import ExerciseNavbar from "../components/Exercises/ExerciseNavbar";
 import ExerciseModal from "../components/Exercises/ExerciseModal";
-import ExerciseAddModal from "../components/Exercises/ExerciseAddModal";
-import { useAuth } from "../hooks/useAuth";
 
 const ExerciseManagement = () => {
-  const { user } = useAuth();
   const [exercises, setExercises] = useState([]);
   const [selectedExercise, setSelectedExercise] = useState({});
   const [isLoading, setIsLoading] = useState(false);
   const [error, setError] = useState("");
   const [selectedCardId, setSelectedCardId] = useState(null);
 
-  const [showAddModal, setShowAddModal] = useState(false);
-  const handleShowAddModal = () => setShowAddModal(true);
 
   const handleCardSelect = (exerciseId) => {
     // Toggle selection: If the clicked card is already selected, unselect it. Otherwise, select it.
@@ -100,35 +95,10 @@ const ExerciseManagement = () => {
   const handleChange = (e) => {
     setSelectedExercise({ ...selectedExercise, [e.target.name]: e.target.value });
   };
-
-  const handleSave = async (e) => {
-    e.preventDefault();
-    const formData = new FormData(e.target);
-    const newExerciseData = Object.fromEntries(formData.entries());
-    // Add the user_who_created_it field to the new exercise data
-    newExerciseData.user_who_created_it = user.user_id;
-    try {
-      const response = await fetch("http://localhost:3500/add_exercise", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify(newExerciseData),
-        credentials: "include",
-      });
-      if (!response.ok) throw new Error("Failed to add new exercise");
-  
-      const addedExercise = await response.json(); // Assuming the server responds with the added exercise data
-      setExercises([...exercises, addedExercise]);
-  
-      setShowAddModal(false);
-      alert("New exercise added successfully!");
-    } catch (err) {
-      alert(err.message);
-    }
-  };
   
   return (
     <div>
-      <ExerciseNavbar onAddExercise={handleShowAddModal} />
+      <ExerciseNavbar />
       <Container className="mt-4">
         {isLoading ? (
           <p>Loading exercises...</p>
@@ -159,8 +129,6 @@ const ExerciseManagement = () => {
           handleSubmit={handleSubmit}
           setModalMode={setModalMode}
         />
-
-        <ExerciseAddModal show={showAddModal} handleClose={() => setShowAddModal(false)} handleSave={handleSave} />
       </Container>
     </div>
   );
